Change the return type of Parametric1D.fitshgo
The return type has changed with the intention of providing an easy way
for the user to tabulate the outcomes of multiple distinct uses of
fitshgo. Should they choose to, the user may concatenate the returned
pandas Series types together (perhaps with evaluated metric scores) and
construct a larger table across a number of different inputs (or test
cases) for later analysis.