Implement SNR vs metric datasheet methods and plot
This commit introduces a core plot that will assist in the development
of fitting procedures: the SNR boxplot, implemented under snr_boxplot.

For a user defined metric, snr_boxplot will produce a series of boxplots
for different levels of specified signal to noise ratio. This allows the
user to easily observe fitting method performance over the entire
parameter space and observe how it is affected by signal to noise ratio.