Introduce demicmate_by_derivative for subsampling
I originally developed this algorithm in the qdetective project but I
have moved it here as I think it has more general application. It has
not been exhaustively tested by any means so I have tagged it as
experimental.

The problem it solvees is providing a method to reduce the number of
data points used in a fitting procedure. For resonance data (which I
work on) you collect many thousands of points, and those off resonance
and have a flat frequency response play a less important role in the
fitting than the actual resonance does. By intelligently down sampling
the data so that these flat features are sparsely sampled and the
narrow resonance itself is densely sampled, we can speed up the fitting
routine without hurting the quality of parameter estimation. I have no
doubt that this problem can arise elsewhere and this technique may help
.