Generalise Parametric1D.fitshgo -> Parametric1D.fit
Prior to this commit the user was restricted to using the scipy shgo
global optimizer. The commit refactors the fitshgo function to allow
the user to select between all the local optimisers and 'shgo',
'dual_annealing' and 'differential_evolution' from scipy's family of
global optimizers. The function call has changed to allow for the
passing of keyword arguments to any of these optimizers and it is up
to the user to ensure their options apply to the specified 'method'.