Pass metadata into metric function
This allows for the sampler to pass down the parameter values used to
generate the sample, into the metric. Thus, the metric can evaluate
how well the estimated paramter matches the `true` parameter without
relying on the state of the Parametric1D object (as was the case before
this commit).