Refuse to set parameters outside bounds for fit
Sometimes the optimiser will converge on a solution outside of the
specified bounds. This usually occurs when the optimizer fails to
enter the sample space at all (and the errf just returns infinity).
Some additional checks were added to limit the parameter values to
the closest bound and throw a more understandable exception if the
parameter value comes back as nan.