Improve pint compatability but remove dependancy
This commit removes the pint dependancy from fitkit. However, Signal1D
and Parametric1D types still have the same compatability with pint
(i.e. the x-axis can be defined with pint types), and additional
compatability, whereby parameters may be defined using pint types and
attempts to substitute them into an expression with unphysical units
will result in an exception.

Because pint's UnitRegistry is not global in scope, it is dangerous to
use one within fitkit when it is expected that the use will define and
use their own within other projects. Pint types from different unit
registries are incompatable. As such I have removed the dependancy of
pint from this project, expecting that the user will define their own
pint types if the wish to use them, or go for a unitless approach.