Evalute sympy.lambdify once outside of __call__
This call is really expensive and was being performed each time the
parametric object was being evaluated. Once the expression is defined it
shouldn't change, only the parameters can change. Therefore we only need
to evaluate lambdify once as soon as the expression is defined. This
commit implements this change and speeds up __call__ by an order of
magnitude.