Implement set method with clipping option
It's annoying to have an error flare up when a parameter value is set or
tested that falls outside the specified boundaries. This set method
provides a clip option which will 'clip' the parameter value to between
the bounds if an attempt is made to set it outside this range. Setting
clip to False gives the default behaviour of raising an exception.