Implement parameter freezing
Sometimes you want to freeze a parameter, preventing the user from
chaning it in the gui-fitter or the optimiser from changing it. In my
context this might be some experimental parameter that we set. I want it
to vary in my model but I know the value and don't want to fit it.
Calling freeze on a parameter name (or list of parameter names) will
freeze these parameters. An unfreeze method is also implemented to
unfreeze previously frozen parameters.