Fixes repeated frozen parameter bug
If the user accidentally freezes the same parameter twice it should only
be added to the list once. This also makes sure that no error occurs if
the user trys to unfreeze a parameter that isn't currently frozen,
allowing for `pm.unfreeze([p for p in pm.v])` type usage.