Round initial parameter value to grid
I found that it's super annoying when the grid is somewhat coarse or
there is a lot of sensitivity to one parameter, and as soon as you move
a slider and then move it back to the initial value you get a different
fit. I've changed how this behaves so that what's displayed always
corresponds to a value you can reach using the sliders. If the rounding
to the slider grid changes the value by more than 1% of the parameter
range then a warning message is given.

The warning message should prompt the user to either decrease the
parameter range or increase the slider resolution. They might decide on
either depending on the context of the fitting.