Remove Signal1D -> breaks decimate (to be fixed)
It was annoying requiring users to conform to the Signal1D datatype as
measured data is typically stored in pandas or numpy arrays.
call_types's type_factory is used to handle the different input types.
We remove Signal1D in this commit in favour of the user using their
preffered types as opposed to converting everything.

Removing Signal1D temporarily breaks decimate.py, which will be fixed in
a later commit.