Implement fitting using scipy.optimize.curve_fit
Curve fit is a nice and fast method that I'm using more and more
frequently. Here we add it as a seperate optimizer option for the user.
Unforunately it has a somewhat different interface to the other
optimizers where it can only fit real functions and returns different
datatypes. We may want to consider calling leastsq directly (as is done
in curve_fit under the hood) for consistency with the other methods.

This commit introduces an additional unit test (requires plottage).